Prevent multiple attacks from one mouse-click
Multiple attacks were called due to the per-frame input updates.
(this cause the animation to appear to shoot twice)
(tho in reality, it was shooting more than 5 times)

Finally went and used the isAttacking (renamed from attacking) bool that was just sitting there. This implementation should make it only fire once during the entirety of the animation.